addressing more comments + some cleanup
