+++ app/build.gradle
@@ -49,7 +49,6 @@ android {
 }
 
 dependencies {
-    implementation project(path: ':data:settings')
     // Compose
     def composeBom = platform('androidx.compose:compose-bom:2022.12.00')
     implementation composeBom
@@ -90,10 +89,13 @@ dependencies {
     def nav_version = "2.5.3"
     implementation "androidx.navigation:navigation-compose:$nav_version"
 
+    // Access Settings data
+    implementation project(path: ':data:settings')
+
     // Camera Preview
     implementation(project(":feature:preview"))
 
-    // Settings
+    // Settings Screen
     implementation(project(":feature:settings"))
 }
