+++ data/settings/src/main/java/test/FakeJcaSettingsSerializer.kt
@@ -16,11 +16,8 @@
 
 package test
 
-import android.content.Context
 import androidx.datastore.core.CorruptionException
-import androidx.datastore.core.DataStore
 import androidx.datastore.core.Serializer
-import androidx.datastore.dataStore
 import com.google.jetpackcamera.settings.DarkModeProto
 import com.google.jetpackcamera.settings.JcaSettings
 import com.google.protobuf.InvalidProtocolBufferException
@@ -28,26 +25,10 @@ import java.io.IOException
 import java.io.InputStream
 import java.io.OutputStream
 
-/*
- * Copyright (C) 2023 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
 
-
-class FakeJcaSettingsSerializer  (
+class FakeJcaSettingsSerializer(
     var failReadWithCorruptionException: Boolean = false
-        ): Serializer<JcaSettings> {
+) : Serializer<JcaSettings> {
 
     override val defaultValue: JcaSettings = JcaSettings.newBuilder()
         .setDarkModeStatus(DarkModeProto.DARK_MODE_SYSTEM)
