+++ feature/settings/build.gradle
@@ -39,8 +39,6 @@ android {
 }
 
 dependencies {
-    implementation project(path: ':domain:camera')
-    implementation 'androidx.test:core-ktx:1.5.0'
     // Compose
     def composeBom = platform('androidx.compose:compose-bom:2022.12.00')
     implementation composeBom
@@ -65,6 +63,8 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-core:3.4.0'
     testImplementation "org.mockito:mockito-core:5.2.0"
     testImplementation "org.jetbrains.kotlinx:kotlinx-coroutines-test:1.6"
+    implementation 'androidx.test:core-ktx:1.5.0'
+
 
     // Guava
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-guava:1.4.1'
